Switch to 64-bit hash function in maps

+++ core/src/main/java/io/questdb/cairo/map/ShardedMapCursor.java
@@ -284,7 +284,7 @@ public class ShardedMapCursor implements MapRecordCursor {
         }
 
         @Override
-        public int keyHashCode() {
+        public long keyHashCode() {
             return baseRecord.keyHashCode();
         }
